requirements.txt
fastapi
uvicorn
redis
asyncpg
python-decouple
bcrypt==4.0.1

passlib
pyjwt
